Rename all files called 'aux.*' to something else; Windows can't handle them.
