+++ util/ego/bo/bo.c
@@ -24,7 +24,7 @@
 #include "../share/lset.h"
 #include "../share/map.h"
 #include "../share/alloc.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "../share/def.h"
 #include "../share/go.h"
