+++ util/ego/cj/cj.c
@@ -22,7 +22,7 @@
 #include "../share/lset.h"
 #include "../share/map.h"
 #include "../share/alloc.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "../share/def.h"
 #include "../share/stack_chg.h"
 #include "../share/go.h"
