+++ util/ego/cs/cs_elim.c
@@ -8,7 +8,7 @@
 #include "../share/types.h"
 #include "../share/alloc.h"
 #include "../share/lset.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "../share/global.h"
 #include "../share/debug.h"
 #include "cs.h"
