+++ util/ego/cs/cs_getent.c
@@ -5,7 +5,7 @@
  */
 #include <em_mnem.h>
 #include "../share/types.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "../share/debug.h"
 #include "../share/global.h"
 #include "cs.h"
