+++ util/ego/cs/cs_profit.c
@@ -10,7 +10,7 @@
 #include "../share/types.h"
 #include "../share/debug.h"
 #include "../share/global.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "../share/cset.h"
 #include "../share/lset.h"
 #include "cs.h"
