+++ util/ego/cs/cs_vnm.c
@@ -10,7 +10,7 @@
 #include "../share/types.h"
 #include "../share/global.h"
 #include "../share/debug.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "cs.h"
 #include "cs_alloc.h"
 #include "cs_aux.h"
