+++ util/ego/ic/ic_aux.c
@@ -17,7 +17,7 @@
 #include "../share/global.h"
 #include "../share/debug.h"
 #include "../share/def.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "ic.h"
 #include "ic_io.h"
 #include "ic_lookup.h"
