+++ util/ego/lv/lv.c
@@ -25,7 +25,7 @@
 #include "../share/map.h"
 #include "../share/get.h"
 #include "../share/put.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "../share/init_glob.h"
 #include "../share/locals.h"
 #include "../share/go.h"
