+++ util/ego/share/utils.h
