+++ util/ego/sr/sr_aux.c
@@ -17,7 +17,7 @@
 #include "../share/debug.h"
 #include "../share/global.h"
 #include "../share/lset.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "sr_aux.h"
 #include "sr_xform.h"
