+++ util/ego/ud/ud_copy.c
@@ -16,7 +16,7 @@
 #include "../share/lset.h"
 #include "../share/cset.h"
 #include "../share/def.h"
-#include "../share/aux.h"
+#include "../share/utils.h"
 #include "../share/locals.h"
 #include "../ud/ud_defs.h"
 #include "ud_copy.h"
